Refactor data processing: replace hf_data_processor with math_hf_data_processor in GRPO examples and tests. Update RewardModelEnvironment configuration to embed dynamic batching and sequence packing settings to code got RM env.
