+++ tests/unit/data/test_data_processor.py
@@ -23,7 +23,6 @@ from datasets import Dataset
 abspath = os.path.abspath(__file__)
 sys.path.append("/".join(abspath.split("/")[:-4]))
 
-from examples.run_grpo_math import hf_data_processor
 from nemo_rl.algorithms.utils import get_tokenizer
 from nemo_rl.data.datasets import AllTaskProcessedDataset
 from nemo_rl.data.datasets.eval_datasets import (
@@ -37,7 +36,7 @@ from nemo_rl.data.datasets.response_datasets import (
     OpenMathInstruct2Dataset,
 )
 from nemo_rl.data.interfaces import TaskDataProcessFnCallable, TaskDataSpec
-from nemo_rl.data.processors import math_data_processor
+from nemo_rl.data.processors import math_data_processor, math_hf_data_processor
 from nemo_rl.models.policy import TokenizerConfig
 
 
@@ -113,9 +112,9 @@ def test_math_hf_data_processor(tokenizer_name, dataset_cls):
     )
 
     task_data_processors: dict[str, tuple[TaskDataSpec, TaskDataProcessFnCallable]] = (
-        defaultdict(lambda: (math_task_spec, hf_data_processor))
+        defaultdict(lambda: (math_task_spec, math_hf_data_processor))
     )
-    task_data_processors["math"] = (math_task_spec, hf_data_processor)
+    task_data_processors["math"] = (math_task_spec, math_hf_data_processor)
 
     dataset = AllTaskProcessedDataset(
         dataset=data.formatted_ds["train"],
